feat: add container, btn component style, all procedure components and procedure context in post page
