+++ src/Components/Posts/DepositAndRent.tsx
@@ -1,0 +1,7 @@
+export default function DepositAndRent() {
+  return (
+    <div>
+      <h3>訂金與租金</h3>
+    </div>
+  );
+}
