+++ src/Components/Posts/Equipments.tsx
@@ -1,0 +1,7 @@
+export default function Equipments() {
+  return (
+    <div>
+      <h3>設備</h3>
+    </div>
+  );
+}
