+++ src/Components/Posts/Features.tsx
@@ -1,0 +1,7 @@
+export default function Features() {
+  return (
+    <div>
+      <h3>重點特色</h3>
+    </div>
+  );
+}
