+++ src/Components/Posts/Introduction.tsx
@@ -1,0 +1,7 @@
+export default function Introduction() {
+  return (
+    <div>
+      <h3>房源介紹</h3>
+    </div>
+  );
+}
