+++ src/Components/Posts/OtherExpenses.tsx
@@ -1,0 +1,7 @@
+export default function OtherExpenses() {
+  return (
+    <div>
+      <h3>雜支</h3>
+    </div>
+  );
+}
