+++ src/Components/Posts/Photos.tsx
@@ -1,0 +1,7 @@
+export default function Photos() {
+  return (
+    <div>
+      <h3>房源照片</h3>
+    </div>
+  );
+}
