+++ src/Components/Posts/Restrictions.tsx
@@ -1,0 +1,7 @@
+export default function Restrictions() {
+  return (
+    <div>
+      <h3>租客限制</h3>
+    </div>
+  );
+}
