+++ src/index.css
@@ -16,12 +16,15 @@
     @apply w-8 h-8 text-center leading-8 bg-gray-400 text-gray-600 rounded-full
   }
   .add-new-input {
-    @apply border border-black rounded-lg px-4 py-3;
+    @apply w-full border border-black rounded-lg px-4 py-3;
   }
   .add-new-input-block {
     @apply flex-1 flex flex-col gap-2;
   }
   .add-new-input-unit {
-    @apply relative after:w-auto after:h-auto after:absolute after:right-3 after:bottom-3 after:z-10;
+    @apply relative after:absolute after:right-3 after:bottom-3 after:z-10;
+  }
+  .add-new-btn {
+    @apply self-start mt-12 px-4 py-2 bg-slate-700 rounded-3xl text-white;
   }
 }
